Extract movement fnality into new component
We do this to separate the intensive map rendering stuff from the easy
movement shifting stuff. This means we can re-render the map only when
we need to, thus saving a lot of time.

This means we can re-jig the UIObserver to grab the current state of the
controls and send it to our store right after we finish the moving animation —
which is now done in JS to give us more predictability (which might not be
necessary now I think of it, but eh).

This puts us in a `keypress -> move -> animate -> move -> animate -> ...`
cycle, giving us smooth walking. We couldn't do this before because re-
rendering the map each time gave us a big pause.

Make sense? I hope so!