Get movement working as hoped (finally)
Still needs refactoring.

I worked out that the logic that decides movement needs to know too much
about the state, so it has to go within redux's world. So it's moved there.
There is a bunch of duplication in here that needs fixing, but I wanted
to save it here while it actually worked!