WIP implementation of abstraction between game & store
This is roughly analogous to a smarter action creator / middleware
layer — which it may hopefully end up being in the final instance.
For now, it solves our problems setting up timed sequences of events
by having promises that are resolved based on the state.

We've extracted a lot of logic from the reducer into the Sequencer layer,
which is good — but we need to be careful it doesn't become a mess.

Things that would be nice:
* Have the sequencer use our Redux store as the backend and thus be
  a pure function. It would be nice but I don't see how yet.
* Normalise all our actions that we're spewing all over the place at
  the moment.
* Break up the `movementIntent` event somehow. It's natural that in
  a game we have an awful lot of fnality attached to very few events,
  but it's not exactly nice.
* Things like `faceDirection(Game.faceMovementDirection(movement))` can
  probably become `faceDirection(Game.movementDirection(movement))` —
  we're effectively duplicating our action creators at the moment.