Change state to state stack model to make portal movement work
Instead of having a single gameState we have a stack of gameStates. So
if someone starts walking the state is:

```
WALKING   <-- head
STANDING
```

So when they stop walking, we pop WALKING off and get STANDING. And when
they're walking into a portal, we have:

```
WALKING
TELEPORTING
STANDING
```

So when the walking anim stops we pop it off, continue teleporting, and
then pop that off to get standing again.

You can see this helps us deal with certain kinds of sequences a
bit better.